requirements.txt
Flask==2.0.2
torchvision==0.12.0
pytorch-lightning==1.6.3
wandb==0.12.16
lightning-bolts==0.5.0
